+++ R/psa.R
@@ -136,11 +136,11 @@ summary.psa <- function(object, calc_sds = FALSE, ...) {
 #'
 #' @importFrom ellipse ellipse
 #' @import dplyr
-#' @import reshape2
 #' @import ggplot2
 #' @importFrom scales dollar_format
 #' @return A \code{ggplot2} plot of the PSA, showing the distribution of each PSA sample and strategy
 #' on the cost-effectiveness plane.
+#' @importFrom tidyr pivot_longer
 #' @export
 plot.psa <- function(x,
                      center = TRUE, ellipse = TRUE,
@@ -160,18 +160,23 @@ plot.psa <- function(x,
   # expect that effectiveness and costs have strategy column names
   # removes confusing 'No id variables; using all as measure variables'
   df_cost <- suppressMessages(
-    melt(cost, variable.name = "Strategy",
-         factorsAsStrings = TRUE,
-         value.name = "Cost")
+    pivot_longer(cost,
+                 everything(),
+                 names_to = "Strategy",
+                 values_to = "Cost")
   )
   df_effect <- suppressMessages(
-    melt(effectiveness, variable.name = "Strategy",
-         factorsAsStrings = TRUE,
-         value.name = "Effectiveness")
+    pivot_longer(effectiveness,
+                 cols = everything(),
+                 names_to = "Strategy",
+                 values_to = "Effectiveness")
   )
   ce_df <- data.frame("Strategy" = df_cost$Strategy,
                       "Cost" = df_cost$Cost,
                       "Effectiveness" = df_effect$Effectiveness)
+
+  # make strategies in psa object into ordered factors
+  ce_df$Strategy <- factor(ce_df$Strategy, levels = strategies, ordered = TRUE)
 
   psa_plot <- ggplot(ce_df, aes_string(x = "Effectiveness", y = "Cost", color = "Strategy")) +
     geom_point(size = 0.7, alpha = alpha, shape = 21) +
@@ -183,6 +188,8 @@ plot.psa <- function(x,
       group_by(.data$Strategy) %>%
       summarize(Cost.mean = mean(.data$Cost),
                 Eff.mean = mean(.data$Effectiveness))
+    # make strategies in psa object into ordered factors
+    strat_means$Strategy <- factor(strat_means$Strategy, levels = strategies, ordered = TRUE)
     psa_plot <- psa_plot +
       geom_point(data = strat_means,
                  aes_string(x = "Eff.mean", y = "Cost.mean", fill = "Strategy"),